feat: Add preferences for location 1 and location 2
This commit adds preferences for location 1 and location 2 in the package.json file. These preferences allow the user to set the location of the database for the respective locations. The preferences include a description, name, title, type, default value, label, and required flag.

The get1.jsx and get2.jsx files have been updated to use the preference values for the API request URLs. The sendrr1.jsx and sendrr2.jsx files have also been updated to use the preference values for the API request URLs.

This change improves the flexibility and customization options for the application.